pyproject.toml
[tool.poetry]
name = "rl-papers"
version = "0.1.0"
description = ""
authors = ["Fraser Mince <frasermince@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7, <3.10"
torch = "^1.9.0"
pytorch-lightning = "^1.6.3"
numpy = "^1.20.2"
lightning-bolts = "^0.3.3"
tensorboard = "^2.4.0"
PyVirtualDisplay = "^2.1"
wandb = "^0.10.27"
torchvision = "^0.12.0"
sklearn = "^0.0"
ipdb = "^0.13.7"
ale-py = "0.7.5"
opencv-python = "^4.5.5.64"
gym = {extras = ["atari", "accept-rom-license"], version = "^0.23.0"}
packaging = "^20.9"
jax = "^0.3.13"
jaxlib = "^0.3.10"
flax = "^0.5.2"
chex = "^0.1.3"
jaxline = "^0.0.5"
absl-py = "^1.1.0"
tensorflow = "^2.9.1"
ml-collections = "^0.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
